Modify CoMap schema to no longer extend `ZodObject`
