+++ packages/jazz-tools/src/tools/tests/zod.test-d.ts
@@ -15,6 +15,21 @@ describe("CoValue and Zod schema compatibility", () => {
     });
   });
 
+  test("cannot use z.discriminatedUnion with CoValue schemas as values", () => {
+    const Dog = co.map({
+      type: z.literal("dog"),
+    });
+
+    const Cat = co.map({
+      type: z.literal("cat"),
+    });
+
+    const Person = co.map({
+      // @ts-expect-error: cannot use z.discriminatedUnion with a CoValue schema
+      pets: z.discriminatedUnion("type", [Dog, Cat]),
+    });
+  });
+
   test("cannot use z.union with CoValue schemas as values", () => {
     const Dog = co.map({
       type: z.literal("dog"),
